+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KQVKDCKD4EPZWYAX4068J66G"
+        "01KQXRQZ7QYJB83ZZPX3Y2KDG6"
     )
     r"""The unique identifier of the data source"""