+++ packages/orq-rc/src/orq_ai_sdk/models/post_v2_feedback_evaluationop.py
@@ -55,7 +55,7 @@ class RequestBody3(BaseModel):
 
     source: Optional[PostV2FeedbackEvaluationRequestBodyRequestSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2026-05-05T08:17:30.669Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2026-05-06T04:29:08.966Z")
     r"""The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -121,7 +121,7 @@ class RequestBody2(BaseModel):
 
     source: Optional[PostV2FeedbackEvaluationRequestBodySource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2026-05-05T08:17:30.668Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2026-05-06T04:29:08.966Z")
     r"""The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -187,7 +187,7 @@ class RequestBody1(BaseModel):
 
     source: Optional[RequestBodySource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2026-05-05T08:17:30.668Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2026-05-06T04:29:08.965Z")
     r"""The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")