+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasetop.py
@@ -154,7 +154,7 @@ class UpdateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-05-05T08:17:15.795Z")
+    updated: Optional[datetime] = parse_datetime("2026-05-06T04:28:53.453Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")