+++ packages/orq-rc/RELEASES.md
@@ -7569,3 +7569,13 @@ Based on:
 - [python v4.9.0-rc.24] packages/orq-rc
 ### Releases
 - [PyPI v4.9.0-rc.24] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.24 - packages/orq-rc
+
+## 2026-05-06 04:32:07
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.762.0 (2.882.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.9.0-rc.25] packages/orq-rc
+### Releases
+- [PyPI v4.9.0-rc.25] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.25 - packages/orq-rc